700_release

+++ gcc/testsuite/gcc.dg/graphite/scop-dsyr2k.c
@@ -1,9 +1,8 @@
-/* { dg-require-effective-target size32plus } */
 #define NMAX 3000
 
 static double a[NMAX][NMAX], b[NMAX][NMAX], c[NMAX][NMAX];
 
-void dsyr2k(int N) {
+void dsyr2k(long N) {
 	int i,j,k;
 	
 #pragma scop
@@ -17,4 +16,6 @@ void dsyr2k(int N) {
 #pragma endscop
 }
 
-/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite" { xfail *-*-* } } } */ 
+/* { dg-final { scan-tree-dump-times "number of SCoPs: 1" 1 "graphite"} } */ 
+/* { dg-final { cleanup-tree-dump "graphite" } } */
+
