+++ gcc/testsuite/gcc.dg/vect/pr31699.c
@@ -1,4 +1,4 @@
-/* { dg-require-effective-target vect_float } */
+/* { dg-require-effective-target vect_double } */
 
 #include <stdlib.h>
 #include <stdarg.h>
@@ -7,9 +7,9 @@
 float x[256];
 
 __attribute__ ((noinline))
-float *foo(void)
+double *foo(void)
 {
- float *z = malloc (sizeof(float) * 256);
+ double *z = malloc (sizeof(double) * 256);
 
  int i;
  for (i=0; i<256; ++i)
@@ -36,3 +36,4 @@ int main()
 /* { dg-final { scan-tree-dump-times "vectorized 1 loops" 2 "vect" { target vect_intfloat_cvt } } } */
 /* { dg-final { scan-tree-dump-times "vector alignment may not be reachable" 1 "vect" { target { ! vector_alignment_reachable } } } } */
 /* { dg-final { scan-tree-dump-times "Alignment of access forced using versioning" 1 "vect" { target { ! vector_alignment_reachable } } } } */
+/* { dg-final { cleanup-tree-dump "vect" } } */
